requirements.txt
torch
einops
deeplake
accelerate
pydantic
pyyaml
transformers
bitsandbytes
git+https://github.com/huggingface/peft  # peft bleeding edge version need for kbit training
smart_open
scipy
torchvision
